feat(tmdb integration): add external layer and fix architecture issues and design

+++ watchly-api/src/main/java/ua/dragunov/watchlyapi/dto/MediaItemPreviewResponse.java
@@ -5,5 +5,5 @@ public record MediaItemPreviewResponse(
         String title,
         int releaseYear,
         String poster,
-        int rating
+        double rating
 ) {}
